requirements.txt
aiogram==3.7.0
google-generativeai==0.6.0
python-dotenv==1.0.1
fpdf2==2.7.9